Fix campo prezzi con Double

+++ CH10-Feign-Client-MongoDB/library-sales/src/main/java/it/large/library/sales/model/ClientTotalAmount.java
@@ -2,11 +2,9 @@ package it.large.library.sales.model;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 public class ClientTotalAmount {
 
-    private BigDecimal totalAmount;
+    private Double totalAmount;
 
 }
